Re-organize imports
