+++ tests/api/test_secure_api.py
@@ -1,7 +1,6 @@
 import json
 
-from connexion.apis import FlaskApi
-from connexion.apps import FlaskApp
+from connexion import FlaskApp
 
 
 def test_security_over_inexistent_endpoints(oauth_requests, secure_api_spec_dir):
