+++ tests/test_metrics.py
@@ -2,7 +2,6 @@ import json
 
 import connexion
 import flask
-from connexion.apis import FlaskApi
 from connexion.decorators.metrics import UWSGIMetricsCollector
 from mock import MagicMock
